Extract rack dependency for applications without rack dependency
