+++ sentry-ruby/lib/sentry/rack.rb
@@ -1,4 +1,4 @@
-require 'time'
 require 'rack'
 
 require 'sentry/rack/capture_exceptions'
+require 'sentry/rack/interface'
